+++ backend/modules/demo/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const routes = require('./routes');
+const service = require('./service');
+
+module.exports = { routes, service };